+++ Movie Library Final Project/Movie Library Final Project/Controllers/UserController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using MovieLibrary.Models.Mediatr.MovieCommands;
 using MovieLibrary.Models.Mediatr.UserCommands;
+using MovieLibrary.Models.Models;
 using MovieLibrary.Models.Requests.MovieRequests;
 using MovieLibrary.Models.Requests.UserRequests;
 
@@ -29,6 +30,8 @@ namespace Movie_Library_Final_Project.Controllers
         [HttpGet("Get a Users")]
         public async Task<IActionResult> GetUsers(int userId)
         {
+            var result = await _mediator.Send(new GetUserByIdCommand(userId));
+            if (result == null) return NotFound("User Does not exist");
             return Ok(await _mediator.Send(new GetUserByIdCommand(userId)));
         }
         [ProducesResponseType(StatusCodes.Status200OK)]
@@ -41,12 +44,18 @@ namespace Movie_Library_Final_Project.Controllers
         [HttpPut("Update A User")]
         public async Task<IActionResult> UpdateMovie([FromBody] UpdateUserRequest user)
         {
+            var result = await _mediator.Send(new GetUserByIdCommand(user.UserId));
+            if (result == null) return NotFound("User Does not exist");
             return Ok(await _mediator.Send(new UpdateUserCommand(user)));
         }
         [HttpDelete("Delete a User")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<IActionResult> Delete(int userId)
         {
+            var result = await _mediator.Send(new GetUserByIdCommand(userId));
+            if (result == null) return NotFound("User Does not exist");
+            bool isThereAPlan = await _mediator.Send(new DoesUserHaveSubscriptionCommand(userId));
+            if (isThereAPlan) return BadRequest("Cannot remove an user with a subscription");
             return Ok(await _mediator.Send(new DeleteUserCommand(userId)));
         }
     }
